Created video-09-row-data-extract.py as refactored version of -08

+++ run-extractor.sh
@@ -19,15 +19,20 @@ FILE_DIR="."
 
 TIME_START=$(date)
 SECONDS=0
-python video-08-row-data-extract.py \
+python video-09-row-data-extract.py \
  --input=data/${FILE_STEM}-input.MOV \
  --output=${FILE_DIR}/${FILE_STEM}-output.avi \
 
 TIME_END=$(date +"%T")
 echo "Started $TIME_START, completed $TIME_END, duration ${SECONDS}s"
 
-mv output.txt ${FILE_DIR}/${FILE_STEM}-output.txt
-mv output_errors.txt ${FILE_DIR}/${FILE_STEM}-output_errors.txt
-
+if [[ -f output.txt ]]; then
+	mv output.txt ${FILE_DIR}/${FILE_STEM}-output.txt
+	echo "Created ${FILE_DIR}/${FILE_STEM}-output.txt"
+fi
+if [[ -f output_errors.txt ]]; then
+	mv output_errors.txt ${FILE_DIR}/${FILE_STEM}-output_errors.txt
+	echo "Created ${FILE_DIR}/${FILE_STEM}-output_errors.txt"
+fi
 
 # https://stackoverflow.com/questions/8903239/how-can-i-calculate-time-elapsed-in-a-bash-script
